Refactor and handle old scripts

+++ src/lynx-scoreboard.ts
@@ -4,9 +4,15 @@ import { LynxResults, LynxDirective } from "./lynx-data.interface";
 import { parseLynxPacket } from "./lynx-data-parser";
 import { Socket } from "net";
 
+export enum Protocol {
+    tcp = "tcp",
+    udp = "udp"
+}
+
 interface LynxScoreboardOpts {
-    ip: string;
+    ip?: string;
     port: number;
+    protocol?: Protocol;
 }
 
 type Topic =
@@ -29,25 +35,6 @@ type Subscribers = {
 };
 
 export class LynxScoreboard {
-    static async listen(opts: LynxScoreboardOpts): Promise<LynxScoreboard> {
-        return new Promise((resolve) => {
-            const scoreboard = new LynxScoreboard(opts);
-            scoreboard.subscribe("listening", () => resolve(scoreboard));
-        });
-    }
-
-    public get ip(): string {
-        return this.opts.ip;
-    }
-
-    public get port(): number {
-        return this.opts.port;
-    }
-
-    public get isListening(): boolean {
-        return this._isListening;
-    }
-
     private readonly _socketUDP: dgram.Socket;
     private readonly _serverTCP: net.Server;
     private _isListening = false;
@@ -59,33 +46,56 @@ export class LynxScoreboard {
         stoppedListening: [],
     };
 
-    private partialMessage = "";
-    private fullMessage = "";
+    private message = "";
     private clients: Socket[] = [];
 
-    private constructor(private opts: LynxScoreboardOpts) {
-        this._socketUDP = dgram.createSocket("udp4");
-        this._serverTCP = net.createServer((tcpServer) => {
-            tcpServer.on("data", buffer => {
-                try {
-                    const message = parseLynxPacket(buffer.toString("utf8"));
-                    const topic = message.isDirective ? "directive" : "results";
-                    this._publish(topic, message.data);
-                } catch (err) {
-                    this._publish("error", String(err));
-                }
-            });
-
-            tcpServer.on("error", (err) => {
-                this._publish("error", err.message);
-            });
-        });
-        this._startListening();
+    public get ip(): string {
+        return this.opts?.ip || "0.0.0.0";
     }
 
-    private _startListening(): void {
-    // Start TCP listener
-        this._serverTCP.listen(this.port, this.ip);
+    public get port(): number {
+        return this.opts.port;
+    }
+
+    public get isListening(): boolean {
+        return this._isListening;
+    }
+
+    static async listen(opts: LynxScoreboardOpts): Promise<LynxScoreboard> {
+        return new Promise((resolve) => {
+            const scoreboard = new LynxScoreboard(opts);
+            scoreboard.subscribe("listening", () => resolve(scoreboard));
+        });
+    }
+
+    private constructor(private opts: LynxScoreboardOpts) {
+        if (this.opts?.protocol === Protocol.tcp) {
+            this._serverTCP = net.createServer((tcpServer) => {
+                tcpServer.on("data", buffer => {
+                    try {
+                        const message = parseLynxPacket(buffer.toString("utf8"));
+                        const topic = message.isDirective ? "directive" : "results";
+                        this._publish(topic, message.data);
+                    } catch (err) {
+                        this._publish("error", String(err));
+                    }
+                });
+
+                tcpServer.on("error", (err) => {
+                    this._publish("error", err.message);
+                });
+            });
+            this.startListeningTCP();
+        } else {
+            this._socketUDP = dgram.createSocket({
+                type: "udp4",
+                reuseAddr: true
+            });
+            this.startListeningUDP();
+        }
+    }
+
+    private startListeningUDP(): void {
         // Open UDP socket
         this._socketUDP
             .on("listening", () => {
@@ -96,22 +106,23 @@ export class LynxScoreboard {
                 this._isListening = false;
                 this._publish("stoppedListening", "UDP");
             })
-            .on("message", (buffer) => {
-                // check if datagram ends with trailer.  If not, more data to come
-                if (!buffer.toString("utf8").endsWith("*COMPLETE")) {
-                    this.partialMessage = this.partialMessage + buffer.toString("utf8");
+            .on("message", (buffer, rinfo) => {
+                // check if datagram is missing trailer and default packet size is full.  If not, more data to come
+                // This is a specific case from FinishLynx.  Older scripts don't have trailer, so we qualify with full packet size
+                if (!buffer.toString("utf8").endsWith("*COMPLETE") && rinfo.size === 536) {
+                    this.message = this.message + buffer.toString("utf8");
                 } else {
-                    // Ends with trailer?  This message is complete
-                    this.fullMessage = this.partialMessage + buffer.toString("utf8");
+                    // Ends with trailer or another size?  This message is complete
+                    this.message = this.message + buffer.toString("utf8");
 
                     try {
-                        const packet = parseLynxPacket(this.fullMessage);
+                        const packet = parseLynxPacket(this.message);
                         const topic = packet.isDirective ? "directive" : "results";
                         this._publish(topic, packet.data);
                         // reset the partial message for next datagram
-                        this.partialMessage = "";
+                        this.message = "";
                     } catch (err) {
-                        this.partialMessage = "";
+                        this.message = "";
                         this._publish("error", String(err));
                     }
                 }
@@ -120,6 +131,15 @@ export class LynxScoreboard {
                 this._publish("error", err.message);
             })
             .bind(this.port, this.ip);
+    }
+
+    private startListeningTCP(): void {
+        // Start TCP listener
+        this._serverTCP.listen(this.port, this.ip);
+        this._serverTCP.on("listening", () => {
+            this._isListening = true;
+            this._publish("listening");
+        });
 
         this._serverTCP.on("connection", (socket) => {
             this.clients.push(socket);
@@ -166,18 +186,24 @@ export class LynxScoreboard {
         return this;
     }
 
-    public stopListening(): Promise<void> {
+    public stopListening(protocol?: Protocol): Promise<void> {
         if (!this._isListening) {
             throw "This socket is not listening.";
         }
         return new Promise((resolve) => {
             this.subscribe("stoppedListening", () => resolve);
-            this._socketUDP.close();
+            if (this._socketUDP && !protocol) {
+                this._socketUDP.close();
+            }
+            if (this.clients && protocol === Protocol.tcp) {
+                this.clients.forEach(client => {
+                    client.destroy();
+                });
+            }
             // For tcp, sever connections before closing server
-            this.clients.forEach(client => {
-                client.destroy();
-            });
-            this._serverTCP.close();
+            if (this._serverTCP && protocol === Protocol.tcp) {
+                this._serverTCP.close();
+            }
         });
     }
 }
